Added Frame data structure Did matching between figures but error in B-10 and B-12 in that process.
NEW:
Added relationships which works
Have to run it for A->B, C-># for every problem

CORRECTION: Meant that have to write transformation method to
build transformation matrix from A->B and A->C
Then, we weight which transformation is "easier to gauge"
Then if horizontal, run transformation on C->#
else run on B->#

for each run transformation weigh again, and choose highest weight as answer? kek'd

ADDED:
tranformation calculator which checks the transformations for each element from each figure to another figure and then assigns a score to the element for each figure.
WILL ADD: a method to decide which direction to consider transformation for 2x2
generate transformations from figure B/C to # and then rank most similar in same direction as answer